requirements.txt
viberbot
flask
python-dotenv
python-statemachine
requests
pytest
mypy
dataclasses-json